pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hiverecon"
version = "0.1.0"
description = "AI-powered reconnaissance framework for bug bounty hunting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "HiveRecon Team", email = "hiverecon@example.com"}
]
keywords = ["security", "recon", "bug-bounty", "ai", "langchain", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.1.0",
    "ollama>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.3",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]
pdf = [
    "weasyprint>=60.2",
    "fpdf2>=2.7.7",
]
all = [
    "hiverecon[dev,pdf]",
]

[project.scripts]
hiverecon = "hiverecon.cli:app"

[project.urls]
Homepage = "https://github.com/hiverecon/hiverecon"
Documentation = "https://github.com/hiverecon/hiverecon/docs"
Repository = "https://github.com/hiverecon/hiverecon"

[tool.setuptools.packages.find]
where = ["."]
include = ["hiverecon*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
